parser: Split parsing logic into modules
