+++ parser/src/parser/tests/mod.rs
@@ -1,14 +1,16 @@
 use assert_matches::assert_matches;
 use nom::{
-    bytes::complete::{escaped_transform, is_not},
+    bytes::complete::{escaped_transform, is_not, take_while1},
     combinator::map_res,
-    multi::fold_many1,
+    multi::{fold_many1, separated_list0},
 };
 
-use super::*;
+use super::{expr::simple_expr, *};
 use crate::{
     alloc::String,
-    grammars::{ParseLiteral, Typed, Untyped},
+    grammars::{Grammar, ParseLiteral, Typed, Untyped},
+    spans::Spanned,
+    BinaryOp, Destructure, Expr, Lvalue, SpannedLvalue, UnaryOp,
 };
 
 mod basics;
