+++ parser/src/parser/tests/object.rs
@@ -6,7 +6,10 @@ use nom::Err as NomErr;
 use super::{lsp, sp, FieldGrammar, Literal, ValueType};
 use crate::{
     parser::{
-        expr, fn_def, lvalue, object_destructure, object_expr, simple_expr, statement, Complete,
+        expr::{expr, object_expr, simple_expr},
+        fn_def,
+        lvalue::{lvalue, object_destructure},
+        statement, Complete,
     },
     BinaryOp, Expr, InputSpan, Lvalue, ObjectDestructure, ObjectDestructureField, ObjectExpr,
     Statement,
